backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
python-socketio==5.11.0
python-multipart==0.0.9
pydantic==2.8.0
redis==5.0.8
python-dotenv==1.0.1
aiohttp==3.9.5
aiofiles==23.2.0